Final Quadcopter submission
This commit includes changes to address errors in my initial submission.
Initially, I had made two mistakes and one omission. First, I
misspecified the if-clause that was used to stop learning if the agent
was successful on four out of the last five episodes. Second, I
misspecified the first part of the reward function, which accidentally
caused the reward to decrease as the agent approached the goal. Finally,
I neglected to specify a learning rate for the actor and critic
networks. This commit fixes both errors and addresses the omission.